PCHR-3046: Move role service to CMSDate dir, add factory
