+++ uk.co.compucorp.civicrm.hrcore/tests/phpunit/CRM/HRCore/Service/CMSRole/DrupalRoleServiceTest.php
@@ -1,6 +1,6 @@
 <?php
 
-use CRM_HRCore_Service_DrupalRoleService as DrupalRoleService;
+use CRM_HRCore_CMSData_Role_DrupalRoleService as DrupalRoleService;
 
 /**
  * @group headless
