+++ uk.co.compucorp.civicrm.hrcore/xml/services.xml
@@ -3,9 +3,9 @@
            xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
 
     <services>
-        <service id="drupal_role_service" class="CRM_HRCore_Service_DrupalRoleService"/>
+        <service id="drupal_role_service" class="CRM_HRCore_CMSData_Role_DrupalRoleService"/>
         <service id="drupal_user_service" class="CRM_HRCore_Service_DrupalUserService">
-            <argument id="drupal_role_service"/>
+            <argument id="drupal_role_service" type="service"/>
         </service>
         <service id="civihr_stats_gatherer" class="CRM_HRCore_Service_CiviHRStatsGatherer">
             <argument id="cms_site_info" type="service"/>
